moved command and parser-related funcs into own internal package
